Add ruby-debug dependency... sort of.
I saw this was being required, so I tried adding it as a normal
dependency, but it looks like the gem doesn't fly with Ruby 1.9, so I
made it version dependent. It seems kind of ghetto, but I'm not sure
what the deal is and this seemed the least invasive approach (as opposed
to just ripping it out).